+++ tests/test_dynamodb_full_table_interruptible_sync.py
@@ -3,10 +3,9 @@ import singer
 
 from boto3.dynamodb.types import TypeSerializer
 
-from tap_tester.scenario import (SCENARIOS)
-import tap_tester.connections as connections
-import tap_tester.menagerie as menagerie
-import tap_tester.runner as runner
+from tap_tester import connections
+from tap_tester import menagerie
+from tap_tester import runner
 
 from base import TestDynamoDBBase
 
@@ -129,6 +128,3 @@ class DynamoDBFullTableInterruptible(TestDynamoDBBase):
             self.assertIsNone(state['bookmarks'][table_name].get('last_evaluated_key'))
 
             self.assertTrue(state['bookmarks'][table_name].get('initial_full_table_complete', False))
-
-
-SCENARIOS.add(DynamoDBFullTableInterruptible)