+++ tests/test_dynamodb_log_based_interruptible.py
@@ -2,10 +2,9 @@ import singer
 
 from boto3.dynamodb.types import TypeSerializer
 
-from tap_tester.scenario import (SCENARIOS)
-import tap_tester.connections as connections
-import tap_tester.menagerie   as menagerie
-import tap_tester.runner      as runner
+from tap_tester import connections
+from tap_tester import menagerie
+from tap_tester import runner
 
 from base import TestDynamoDBBase
 
@@ -202,7 +201,3 @@ class DynamoDBLogBased(TestDynamoDBBase):
             # as the full table sync
             state['bookmarks'][table_name].pop('finished_shards')
             menagerie.set_state(conn_id, state, version=state_version)
-
-
-
-SCENARIOS.add(DynamoDBLogBased)