+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0
+  * Expression attributes implementation [#37](https://github.com/singer-io/tap-dynamodb/pull/37)
+  * Implement request timeout [#38](https://github.com/singer-io/tap-dynamodb/pull/38)
+  * Missing parent objects from projection expression [#38](https://github.com/singer-io/tap-dynamodb/pull/37)
+
+## 1.1.2
+  * Fix issue where decimals can throw a Rounded signal if they are too wide [#33](https://github.com/singer-io/tap-dynamodb/pull/33)
+
 ## 1.1.1
   * Add more error checking to ensure a projection provided is also not empty [#26](https://github.com/singer-io/tap-dynamodb/pull/26)